+++ show.c
@@ -271,7 +271,7 @@ ShowRollout ( const rolloutcontext *prc ) {
             _("Rollout as opening move enabled.") :
             _("Rollout as opening move disabled.") );
   outputf ( _("%s dice generator with seed %u.\n"),
-            gettext ( aszRNG[ prc->rngRollout ] ), prc->nSeed );
+            gettext ( aszRNG[ prc->rngRollout ] ), (unsigned int)prc->nSeed );
             
   /* see if the players settings are the same */
   if (EvalCmp (prc->aecChequer, prc->aecChequer + 1, 1) ||
